compute.absCN {absCNseq}	R Documentation

Convert copy ratios into abosolute copy numbers when given the purity and ploidy estimates

Description

This function converts copy ratios into abosolute copy numbers when the tumor purity and ploidy are provided by the user. The tumor purity and ploidy parameter should be taken from the result after running the "run.absCNSeq" function.

Usage

compute.absCN(seg.data, alpha.opt, tau.opt, qmax = 7)

Arguments

seg.data
a data frame with five columns: "chrom", "loc.start", "loc.end", "eff.seg.len", "normalized.ratio".

alpha.opt
tumor purity estimate

tau.opt
tumor ploidy estimate

qmax
maximum allowed absolute copy number for any segments

Value

The input data frame is augmented with two additional columns: rhat (expected copy ratio) and CN (absolute copy number)

Author(s)

Lei Bao

Examples

library(absCNseq)
my.res.list <- run.absCNSeq("example.cn.txt", "example.snv.txt", "myResult", "Sample1", seq.type="WES", min.seg.len=200)
seg.CN <- compute.absCN(my.res.list$seg.dat, my.res.list$searchRes[i,"alpha"], my.res.list$searchRes[i,"tau"])  # the i-th solution

[Package absCNseq version 1.0 Index]