plot.absCN {absCNseq}	R Documentation

plot estimated integer copy numbers from absCNseq for each chromosome.

Description

plot estimated integer copy numbers from absCNseq for each chromosome.

Usage

plot.absCN(seg.data, chromnum = 1, rawdata = NULL, chromvar = "chrom", loc.start = "chr_start", loc.end = "chr_stop", normalized.ratio = "log2_ratio")

Arguments

seg.data
A result data frame generated by the compute.absCN function. It has five mandatory columns: "chrom", "loc.start", "loc.end", "r", "CN".

chromnum
which chromosome to be plotted. Must be a number from 1 to 22

rawdata
(optional): raw copy ratio data before segmentation. This data frame has four mandatory columns: chromosome number, start position, end position, and normalized copy ratio of tumor vs. normal sample on log2 scale (log2_ratio). The column names are provided by the user. Default column names are taken from the VARSCAN output.

chromvar
The variable name of the chromosome number column for the optional raw data.

loc.start
The variable name of the start position column for the optional raw data.

loc.end
The variable name of the end position column for the optional raw data.

normalized.ratio
The variable name of the log2_ratio column for the optional raw data.

Author(s)

Lei Bao

Examples

library(absCNseq)
my.res.list <- run.absCNSeq("example.cn.txt", "example.snv.txt", "myResult", "Sample1", seq.type="WES", min.seg.len=200)
seg.CN <- compute.absCN(my.res.list$seg.dat, my.res.list$searchRes[1,"alpha"], my.res.list$searchRes[1,"tau"])  # the top first solution
plot.absCN(seg.CN, chromnum=1)  # plot chromosome 1

[Package absCNseq version 1.0 Index]