run.absCNSeq {absCNseq}	R Documentation

main function to run absCNseq

Description

This wrapper function accepts data files and user specified parameters and runs the absCNseq algorithm.

Usage

run.absCNSeq(seg.fn, snv.fn = NULL, res.dir, smp.name, seq.type = c("WES", "WGS"), alpha.min = 0.2, alpha.max = 1, tau.min = 1.5, tau.max = 5, min.freq = 0, min.seg.len = 0, qmax = 7, lamda = 0.5, verbose = FALSE)

Arguments

seg.fn
The name of the file containing the segmentation data. An example file called "example.cn.txt" can be found under the "example" folder within the tarball.

snv.fn
The name of the file containing a set of somatic single nucleotide variants (SNVs). An example file called "example.snv.txt" can be found under the "example" folder within the tarball.

res.dir
The output directory

smp.name
Sample name

seq.type
Either "WES" (whole exome sequencing) or "WGS" (whole genome sequencing)

alpha.min
The minimum allowed value for tumor purity. Default is 0.20. If you do have the pathologist estimate, set it as the lower bound of the pathologist estimate is usually preferred.

alpha.max
The maximum allowed value for tumor purity. Default is 1.0. If you do have the pathologist estimate, set it as the upper bound of the pathologist estimate is usually preferred.

tau.min
The minimum allowed value for tumor ploidy

tau.max
The maximum allowed value for tumor ploidy

min.sol.freq
A solution should appear at least this many times to be kept. Singleton solutions are usually not trustable. By default (min.sol.freq=0), the program will only retain solutions that cover at least 1 percent of the search space.

min.seg.len
The minimum length of a segment to be included in computation. The default value is 200 bp for WES and 3000 bp for WGS.

qmax
Maximum allowed absolute copy number for any segments.

lamda
The relative weight of the segment copy ratio data over the SNV data. Must be a value in(0.0,1.0].

verbose

Details

Please refer to the "example" folder to see the format of the input segmentation file or SNV file.

Value

a list is returned

searchRes
a data frame giving the solution set (purity and ploidy pairs) ranked by their fitting errors

absCN
a data frame giving the absolute copy number estimates of tumor cells for the top first solution (purity and ploidy pair)

absSNV
a data frame giving the absolute multiplicity estimates of SNVs for the top first solution (purity and ploidy pair)

orig.seg.dat
original copy ratio data read from the input segmentation file

seg.dat
filtered copy ratio data

orig.snv.dat
original SNV data read from the input SNV file

snv.dat
filtered SNV data

Author(s)

Lei Bao

Examples

library(absCNseq)
my.res.list <- run.absCNSeq("example.cn.txt", "example.snv.txt", "myResult", "Sample1", seq.type="WES", min.seg.len=200)
seg.CN <- compute.absCN(my.res.list$seg.dat, my.res.list$searchRes[i,"alpha"], my.res.list$searchRes[i,"tau"])  # the i-th solution
plot.absCN(seg.CN, chromnum=1)  # plot chromosome 1

[Package absCNseq version 1.0 Index]